chore(fix-code-duplocation): Remove duplicated codes on login and register controllers
