+++ .babelrc
@@ -1,9 +1,9 @@
 {
-"presets": ["es2015", "react", "stage-0"],
+"presets": ["es2015", "stage-0"],
   "env": {
     "test": {
       "plugins": [ "istanbul"],
-      "presets": ["es2015", "react", "stage-0"]
+      "presets": ["es2015", "stage-0"]
     },
     "production": {
       "plugins": ["transform-remove-console"]
