+++ server/controllers/Book.js
@@ -7,8 +7,8 @@ export default {
    * @param  {object} req request
    * @param  {object} res response
    * Route: POST: /books  
+   * @return {Object} - Object containing status code and success message
    */
-
   create(req, res) {
     return Book.create(req.userInput)
       .then(() =>
@@ -16,15 +16,15 @@ export default {
           message: 'Book uploaded successfully'
         })
       )
-      .catch(error => res.status(400).send(error));
+      .catch(error => res.status(500).send(error));
   },
 
   /** User rent book
    * @param  {object} req - request
    * @param  {object} res - response
    * ROUTE: POST: /users/:userId/books
+   * @return {Object} - Success message
    */
-
   rentBook(req, res) {
     const cur = new Date(),
       after30days = cur.setDate(cur.getDate() + 30);
@@ -68,7 +68,7 @@ export default {
           message: 'You have successfully rented the book'
         })
       )
-      .catch(error => res.status(400).send(error));
+      .catch(error => res.status(500).send(error));
   },
 
   /**
@@ -97,8 +97,8 @@ export default {
    * @param  {object} req request
    * @param  {object} res response
    *  Route: GET: /api/books
+   * @return {Object} - Book details
    */
-
   getBooks(req, res) {
     const pageNum = Number(req.query.page);
     let offset;
@@ -117,21 +117,21 @@ export default {
     })
       .then((books) => {
         if (books.count < 1) {
-          res.status(400).send({
+          res.status(404).send({
             message: 'There is no book in the database'
           });
         } else {
           res.status(201).send(books);
         }
       })
-      .catch(error => res.status(404).send(error));
+      .catch(error => res.status(500).send(error));
   },
 
   /** Adds a new category
    * @param  {object} req request
-   * @param  {0bject} res response
+   * @param  {Object} res response
+   * @return {Object} - return lists of category
    */
-
   addCategory(req, res) {
     return Category.create(req.body)
       .then((category) => {
@@ -143,17 +143,19 @@ export default {
         }
       })
       .catch(error =>
-        res.status(401).send({
+        res.status(500).send({
           error
         })
       );
   },
   /** Dislay users rented books
-   * @param  {object} req request
-   * @param  {0bject} res response
+   * @param  {object} req - request
+   * 
+   * @param  {Object} res - response
+   * 
    * Route: GET: //api/users/:UserId/books?returned=false
+   * @return {Object} - Status code and book data
    */
-
   rentedBooks(req, res) {
     return RentedBook.findAll({
       where: {
@@ -163,22 +165,22 @@ export default {
     })
       .then((books) => {
         if (books.length < 1) {
-          res.status(201).send({
+          res.status(200).send({
             message: 'No rented unreturned books'
           });
         } else {
-          res.status(201).send(books);
+          res.status(200).send(books);
         }
       })
-      .catch(error => res.status(404).send(error));
+      .catch(error => res.status(500).send(error));
   },
 
   /** Admin modify book details
-   * @param  {object} request
-   * @param  {object} resonse
+   * @param  {object} req - request
+   * @param  {object}  res -resonse
    * Route: GET: /
+   * @return {Object} - Success message
    */
-
   modifyBook(req, res) {
     return Book.update(req.body, {
       where: {
@@ -187,20 +189,20 @@ export default {
     })
       .then(() => {
         Book.findById(req.params.bookId).then((book) => {
-          res.status(200).send({
+          res.status(201).send({
             book,
             message: 'Book updated successfully!'
           });
         });
       })
-      .catch(error => res.status(400).send(error));
+      .catch(error => res.status(500).send(error));
   },
   /** User get a specific book
    * @param  {Object} req - request
    * @param  {Object} res - response
    * Route: GET: /books/:bookId 
+   * @returns {Object} - status code and book details
    */
-
   getOneBook(req, res) {
     return Book.findAll({
       where: {
@@ -210,14 +212,14 @@ export default {
       .then((books) => {
         res.status(200).send(books);
       })
-      .catch(error => res.status(404).send(error));
+      .catch(error => res.status(500).send(error));
   },
   /** Admin delete a book
-   * @param  {} req - request
-   * @param  {} res - reponse
+   * @param  {Object} req - request
+   * @param  {Object} res - reponse
    * ROute: DELETE: /books/delete/:bookId
+   * @returns {Object} - returns success message
    */
-
   deleteBook(req, res) {
     return Book.destroy({
       where: {
@@ -230,14 +232,14 @@ export default {
           id: req.params.bookId
         });
       })
-      .catch(error => res.status(404).send(error));
+      .catch(error => res.status(500).send(error));
   },
   /** Get rented books by a specific user
    * @param  {Object} req - request
    * @param  {object} res - response
    * Route: GET: /books/logs/:userId
+   * @returns {Object} - return lists of rented book by a user
    */
-
   rentedBookByUser(req, res) {
     return RentedBook.findAll({
       where: {
@@ -253,14 +255,14 @@ export default {
           res.status(200).send(books);
         }
       })
-      .catch(error => res.status(404).send({ message: error }));
+      .catch(error => res.status(500).send({ message: error }));
   },
   /** User return rented book
    * @param  {object} req - request
    * @param  {object} res - response
    * Route: PUT: /users/:userId/books
+   * @return {Object} - return list of rented books
    */
-
   returnBook(req, res) {
     return RentedBook.update(
       {
@@ -300,15 +302,15 @@ export default {
           })
         )
       )
-      .catch(error => res.status(400).send(error));
+      .catch(error => res.status(500).send(error));
   },
 
   /** Gets the list of category from database
    * @param  {object} req - request
    * @param  {object} res - response
    * Route: PUT: /books/category
+   * @return {Object} - Return category from database
    */
-
   getCategory(req, res) {
     return Category.findAll({})
       .then((category) => {
@@ -317,6 +319,13 @@ export default {
       .catch(error => res.status(500).send(error));
   },
 
+  /**
+   * 
+   * Book search controller
+   * @param {Object} req - request
+   * @param {Object} res - response
+   * @returns {Object} - Returns search result
+   */
   search(req, res) {
     return Book.findAll({
       where: {
